Don't bundle html-to-text in production commons.js
