+++ utils/htmlToString.ts
@@ -1,0 +1,11 @@
+import { fromString } from "html-to-text"
+
+export function htmlToPlaintext(html: string): string {
+    return fromString(html, {
+        tables: true,
+        ignoreHref: true,
+        wordwrap: false,
+        uppercaseHeadings: false,
+        ignoreImage: true
+    })
+}
