+++ utils/string.ts
@@ -4,18 +4,6 @@ export function findUrlsInText(s: string): string[] {
     return s.match(URL_REGEX) || []
 }
 
-import { fromString } from "html-to-text"
-
-export function htmlToPlaintext(html: string): string {
-    return fromString(html, {
-        tables: true,
-        ignoreHref: true,
-        wordwrap: false,
-        uppercaseHeadings: false,
-        ignoreImage: true
-    })
-}
-
 export function parseBool(input: string): boolean {
     const normalized = input.trim().toLowerCase()
     return normalized === "true" || normalized === "1"
